refactored end2end tests into functional tests
