+++ tests/at_end2end_test/test/notify_test.dart
@@ -122,61 +122,6 @@ void main() async {
       });
     });
   });
-
-  group('A group of tests for notification fetch', () {
-    test('A test to verify non existent notification', () async {
-      await AtClientManager.getInstance().setCurrentAtSign(
-          currentAtSign,
-          namespace,
-          TestPreferences.getInstance().getPreference(currentAtSign));
-      var notificationResult = await AtClientManager.getInstance()
-          .atClient
-          .notificationService
-          .fetch('abc-123');
-      expect(notificationResult.id, 'abc-123');
-      expect(notificationResult.status, 'NotificationStatus.expired');
-    });
-
-    test('A test to verify the notification expiry', () async {
-      for (int i = 0; i < 10; i++) {
-        print('Testing notification expiry - test run #$i');
-        await AtClientManager.getInstance().setCurrentAtSign(
-            currentAtSign,
-            namespace,
-            TestPreferences.getInstance().getPreference(currentAtSign));
-        var atKey = (AtKey.shared('test-notification-expiry',
-                namespace: 'wavi', sharedBy: currentAtSign)
-              ..sharedWith(sharedWithAtSign))
-            .build();
-        NotificationResult notificationResult =
-            await AtClientManager.getInstance()
-                .atClient
-                .notificationService
-                .notify(NotificationParams.forUpdate(atKey,
-                    notificationExpiry: Duration(minutes: 1)));
-
-        AtNotification atNotification = await AtClientManager.getInstance()
-            .atClient
-            .notificationService
-            .fetch(notificationResult.notificationID);
-
-        var actualExpiresAtInEpochMills = DateTime.fromMillisecondsSinceEpoch(
-                atNotification.expiresAtInEpochMillis!)
-            .toUtc()
-            .millisecondsSinceEpoch;
-        var expectedExpiresAtInEpochMills =
-            DateTime.fromMillisecondsSinceEpoch(atNotification.epochMillis)
-                .toUtc()
-                .add(Duration(minutes: 1))
-                .millisecondsSinceEpoch;
-        expect(
-            (actualExpiresAtInEpochMills - expectedExpiresAtInEpochMills)
-                    .abs() <
-                10,
-            true);
-      }
-    });
-  });
 }
 
 /// Class responsible for getting the notifications from the cloud secondary
