+++ tests/at_functional_test/config/config.yaml
@@ -1,0 +1,10 @@
+# The @Protocol root server configuration
+root_server:
+  # The url to connect to root server
+  url: 'vip.ve.atsign.zone'
+
+# alice atsign details
+atSign:
+  firstAtSign: "@alice🛠"
+  secondAtSign: "@bob🛠"
+
