+++ tests/at_functional_test/test/atclient_putText_test.dart
@@ -1,14 +1,18 @@
+
 import 'package:at_client/at_client.dart';
+import 'package:at_functional_test/src/config_util.dart';
 import 'package:test/test.dart';
 import 'test_utils.dart';
 
 /// The tests verify the put and get functionality where key is created using AtKey
 /// static factory methods
 void main() {
-  late AtClientManager atClientManager;
-  String atSign = '@alice🛠';
-  String namespace = 'wavi';
+ late AtClientManager atClientManager;
+  late String atSign;
+  final namespace = 'wavi';
+
   setUpAll(() async {
+    atSign = ConfigUtil.getYaml()['atSign']['firstAtSign'];
     atClientManager = await TestUtils.initAtClient(atSign, namespace);
   });
