+++ tests/at_functional_test/test/atclient_put_test.dart
@@ -2,6 +2,7 @@ import 'dart:io';
 import 'dart:typed_data';
 
 import 'package:at_client/at_client.dart';
+import 'package:at_functional_test/src/config_util.dart';
 import 'package:test/test.dart';
 import 'test_utils.dart';
 
@@ -9,10 +10,13 @@ import 'test_utils.dart';
 /// class
 void main() {
   late AtClientManager atClientManager;
-  String atSign = '@alice🛠';
-  String namespace = 'wavi';
+  late String atSign;
+  final namespace = 'wavi';
+
   setUpAll(() async {
+    atSign = ConfigUtil.getYaml()['atSign']['firstAtSign'];
     atClientManager = await TestUtils.initAtClient(atSign, namespace);
+    atClientManager.atClient.syncService.sync();
   });
 
   test('put method - create a key sharing to other atSign', () async {
