+++ tests/at_functional_test/test/atclient_remote_secondary_test.dart
@@ -1,15 +1,18 @@
 import 'package:at_client/at_client.dart';
 import 'package:at_commons/src/verb/llookup_verb_builder.dart';
 import 'package:at_commons/src/verb/update_verb_builder.dart';
+import 'package:at_functional_test/src/config_util.dart';
 import 'package:test/test.dart';
 
 import 'test_utils.dart';
 
 void main() {
   late AtClientManager atClientManager;
-  String atSign = '@alice🛠';
-  String namespace = 'wavi';
+  late String atSign;
+  final namespace = 'wavi';
+
   setUpAll(() async {
+    atSign = ConfigUtil.getYaml()['atSign']['firstAtSign'];
     atClientManager = await TestUtils.initAtClient(atSign, namespace);
   });
