+++ tests/at_functional_test/test/enrollment_test.dart
@@ -1,5 +1,6 @@
 import 'dart:convert';
 
+import 'package:at_functional_test/src/config_util.dart';
 import 'package:test/test.dart';
 import 'package:at_client/at_client.dart';
 import 'package:at_demo_data/at_demo_data.dart' as at_demos;
@@ -7,7 +8,7 @@ import 'test_utils.dart';
 
 void main() {
   late AtClientManager atClientManager;
-  String atSign = '@alice🛠';
+  late String atSign;
   String namespace = 'wavi';
   late String aliceApkamSymmetricKey;
   late String aliceDefaultEncryptionPrivateKey;
@@ -15,6 +16,7 @@ void main() {
   late String alicePkamPublicKey;
 
   setUp(() async {
+    atSign = ConfigUtil.getYaml()['atSign']['firstAtSign'];
     atClientManager = await TestUtils.initAtClient(atSign, namespace);
     aliceApkamSymmetricKey = at_demos.apkamSymmetricKeyMap[atSign]!;
     aliceDefaultEncryptionPrivateKey =
